Remove SnapLocalRepository
